+++ backend/packages/observability-otel/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @cat-factory/observability-otel
 
+## 0.18.38
+
+### Patch Changes
+
+- Updated dependencies [a634746]
+  - @cat-factory/contracts@0.297.0
+  - @cat-factory/kernel@0.288.0
+
+## 0.18.37
+
+### Patch Changes
+
+- Updated dependencies [7893f35]
+  - @cat-factory/contracts@0.296.0
+  - @cat-factory/kernel@0.287.0
+
 ## 0.18.36
 
 ### Patch Changes