+++ backend/packages/sandbox-fixtures/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @cat-factory/sandbox-fixtures
 
+## 0.7.337
+
+### Patch Changes
+
+- Updated dependencies [a634746]
+  - @cat-factory/contracts@0.297.0
+
+## 0.7.336
+
+### Patch Changes
+
+- Updated dependencies [7893f35]
+  - @cat-factory/contracts@0.296.0
+
 ## 0.7.335
 
 ### Patch Changes